Some code cleanup: if ... else if... replaced with switch()
Still the ISR timing is bad, hangs too often. Lack of decoupling
is not the problem. Even with split power supply for logic/drivers
and some 2000µF caps it doesn't go away. THE ISR just stops and
resumes after a while. Reset is not detected, at least my
'we just did a reset' LED doesn't blink.