Improvements...
I tried to only let the pin-change interrupt run when the input buffer
was empty to save cpu cycles. Problem: when main had read the buffer
the interrupt would start in the middle of an incoming byte and got
nonsense data. Now every single incoming byte is _always_ sampled, but
only copied to the input buffer when it empty. Takes more cpu time,
but data is now read reliably.